#161 Updating to Youtube API v3
Solution file works with new API, but requires an API key.  The file
currently uses mine, which is fine with me, but feel free to change it.

Instructions are updated with new links and information on obtaining an
API key.  However, Step 5 is actually incorrect.  There was no
instruction in the previous exercise to set up ajax loading of an
external json, and the linked starter file, jsapis_solution.html,
doesn't have it, either.  Linking a new starter file (like
ajax/json_solution.html) might be a bit of a hassle because students
would also have to download the videos.json file.  Might be better to
have a quick slide reviewing the ajax command and modify the Youtube
instructions to expect adding the ajax command to the existing doc.